move dto package inside model
