+++ thryve-backend/src/main/java/bg/sofia/uni/fmi/webjava/backend/model/dto/UpdateUserDto.java
@@ -1,4 +1,4 @@
-package bg.sofia.uni.fmi.webjava.backend.dto;
+package bg.sofia.uni.fmi.webjava.backend.model.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.Size;
